06_게시판/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
whitenoise==6.2.0
